api/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
